requirements.txt
grpcio==1.71.0
grpcio-reflection==1.71.0
grpcio-tools==1.71.0
protobuf==5.29.4
pyirsdk==1.3.5
python-dotenv==1.1.0
PyYAML==6.0.2
setuptools==78.1.0
